requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
diff-match-patch==20241021
Django==5.2.9
django-crispy-forms==2.5
django-filter==25.2
django-import-export==4.3.14
pillow==12.0.0
reportlab==4.4.7
sqlparse==0.5.5
tablib==3.9.0
